+++ src/collective/eeafaceted/collectionwidget/vocabulary.py
@@ -1,5 +1,6 @@
 # encoding: utf-8
 
+from Products.CMFPlone.utils import safe_unicode
 from collective.behavior.talcondition.behavior import ITALCondition
 from collective.behavior.talcondition.interfaces import ITALConditionable
 from collective.eeafaceted.collectionwidget.interfaces import ICollectionCategories
@@ -9,17 +10,16 @@ from eea.facetednavigation.interfaces import IFacetedNavigable
 from imio.helpers.cache import get_cachekey_volatile
 from plone import api
 from plone.memoize import ram
-from Products.CMFPlone.utils import safe_unicode
 from zope.component import getAdapter
 from zope.globalrequest import getRequest
-from zope.interface import implements
+from zope.interface import implementer
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 
 
+@implementer(IVocabularyFactory)
 class CollectionVocabulary(object):
-    implements(IVocabularyFactory)
 
     def __call__(self, context, real_context):
         self.context = context
@@ -109,8 +109,8 @@ class CollectionVocabulary(object):
 CollectionVocabularyFactory = CollectionVocabulary()
 
 
+@implementer(IVocabularyFactory)
 class CollectionCategoryVocabulary(object):
-    implements(IVocabularyFactory)
 
     def __call__(self, context):
         # find root